Feat : seperate WalletInformation component
