+++ src/pages/Staking/index.tsx
@@ -1,11 +1,7 @@
 import {
   Container,
-  Center,
   Box,
   Text,
-  Heading,
-  Button,
-  Grid,
   Flex,
   Link,
   useColorMode,
@@ -15,7 +11,7 @@ import {IconClose} from 'components/Icons/IconClose';
 import {IconOpen} from 'components/Icons/IconOpen';
 import {Head} from 'components/SEO';
 import {useAppDispatch, useAppSelector} from 'hooks/useRedux';
-import React, {FC, Fragment, useCallback, useMemo} from 'react';
+import React, {Fragment, useCallback, useMemo} from 'react';
 import {shortenAddress} from 'utils';
 import {StakingTable} from './StakingTable';
 import {selectStakes} from './staking.reducer';
@@ -31,32 +27,17 @@ import {
   WithdrawalOptionModal,
   SwapModal,
 } from './StakeOptionModal';
+import {WalletInformation} from './components/WalletInformation';
 import {AppDispatch} from 'store';
-import {openModal, closeModal} from 'store/modal.reducer';
 import {ManageModal} from './StakeOptionModal/Manage/index';
 import {formatStartTime} from 'utils/timeStamp';
 import {useState} from 'react';
 import {Stake} from './staking.reducer';
-import {fetchManageModalPayload} from './utils';
-import {ModalType} from 'store/modal.reducer';
-import {LoadingComponent} from 'components/Loading';
-import {
-  getTotalStakers,
-  getUserBalance,
-  getUserTonBalance,
-} from 'client/getUserBalance';
+import {getTotalStakers, getUserBalance} from 'client/getUserBalance';
 //@ts-ignore
 import {Dot} from 'react-animated-dots';
 import {useEffect} from 'react';
-import {closeSale} from './actions';
-
-type WalletInformationProps = {
-  dispatch: AppDispatch;
-  data: Stake;
-  user: User;
-  account: string | undefined;
-};
-
+import {LoadingDots} from 'components/Loader/LoadingDots';
 type GetDateTimeType =
   | 'sale-start'
   | 'sale-end'
@@ -68,16 +49,6 @@ type GetDateProp = {
   currentBlock: number;
   contractAddress: string;
   type: GetDateTimeType;
-};
-
-const LoadingDots = () => {
-  return (
-    <Flex h={30}>
-      <Dot>·</Dot>
-      <Dot>·</Dot>
-      <Dot>·</Dot>
-    </Flex>
-  );
 };
 
 const GetDate = ({time, currentBlock, contractAddress, type}: GetDateProp) => {
@@ -106,236 +77,6 @@ const GetDate = ({time, currentBlock, contractAddress, type}: GetDateProp) => {
         </Text>
       )}
     </>
-  );
-};
-
-const WalletInformation: FC<WalletInformationProps> = ({
-  user,
-  data,
-  dispatch,
-  account,
-}) => {
-  const {colorMode} = useColorMode();
-  const [loading, setLoading] = useState(false);
-  const [userTonBalance, setUserTonBalance] = useState<string>('');
-  const [stakeBalance, setStakeBalance] = useState<string | undefined>(
-    undefined,
-  );
-  const [tosBalance, setTosBalance] = useState<string | undefined>(undefined);
-  const [stakeDisabled, setStakeDisabled] = useState(true);
-  const [unstakeDisabled, setUnstakeDisabled] = useState(true);
-  const [claimDisabled, setClaimDisabled] = useState(true);
-  const [manageDisabled, setManageDisabled] = useState(true);
-  const currentBlock: number = Number(data.fetchBlock);
-  const miningStart: number = Number(data.miningStartTime);
-  const miningEnd: number = Number(data.miningEndTime);
-  const saleStart: number = Number(data.saleStartTime);
-  const endSaleBtnDisabled =
-    account === undefined || miningStart >= currentBlock ? true : false;
-  const manageBtnDisabled =
-    account === undefined || miningEnd <= currentBlock ? true : false;
-
-  const btnDisabledStake = () => {
-    return account === undefined || saleStart >= currentBlock
-      ? setStakeDisabled(true)
-      : setStakeDisabled(false);
-  };
-
-  const btnDisabledUnstake = () => {
-    return account === undefined ||
-      currentBlock <= miningEnd ||
-      stakeBalance === undefined ||
-      stakeBalance === '0.00'
-      ? setUnstakeDisabled(true)
-      : setUnstakeDisabled(false);
-  };
-
-  const btnDisabledClaim = () => {
-    return account === undefined ||
-      data.saleClosed === false ||
-      tosBalance === undefined ||
-      tosBalance === '0.00'
-      ? setClaimDisabled(true)
-      : setClaimDisabled(false);
-  };
-
-  const manageDisableClaim = () => {
-    return account === undefined || data.saleClosed === false
-      ? setManageDisabled(true)
-      : setManageDisabled(false);
-  };
-
-  useEffect(() => {
-    if (user.address !== undefined) {
-      getWalletTonBalance();
-    }
-    btnDisabledStake();
-    btnDisabledUnstake();
-    btnDisabledClaim();
-    manageDisableClaim();
-    /*eslint-disable*/
-  }, [account, data, dispatch, tosBalance]);
-
-  const modalPayload = async (data: any) => {
-    const result = await fetchManageModalPayload(
-      data.library,
-      data.account,
-      data.contractAddress,
-      data.vault,
-    );
-
-    return result;
-  };
-
-  const getWalletTonBalance = async () => {
-    const result = await getUserTonBalance({
-      account: user.address,
-      library: user.library,
-    });
-    const userBalance = await getUserBalance(data.contractAddress);
-
-    if (result && userBalance) {
-      const trimNum = Number(result).toLocaleString(undefined, {
-        minimumFractionDigits: 2,
-      });
-      setUserTonBalance(trimNum);
-      setStakeBalance(userBalance.totalStakedBalance);
-      setTosBalance(userBalance.rewardTosBalance);
-    }
-  };
-
-  const modalData = useCallback(async (modal: ModalType) => {
-    setLoading(true);
-    let payload;
-
-    try {
-      if (modal === 'manage' || modal === 'claim') {
-        const payloadModal = await modalPayload(data);
-        payload = {
-          ...data,
-          ...payloadModal,
-          user,
-        };
-      } else if (modal === 'unstake') {
-        const payloadModal = await getUserBalance(data.contractAddress);
-        payload = {
-          ...data,
-          totalStakedBalance: payloadModal?.totalStakedBalance,
-        };
-      } else {
-        payload = {
-          ...data,
-          user,
-        };
-      }
-    } catch (e) {
-      console.log(e);
-      setLoading(false);
-    }
-
-    setLoading(false);
-    dispatch(openModal({type: modal, data: payload}));
-  }, []); // eslint-disable-line react-hooks/exhaustive-deps
-
-  const theme = useTheme();
-  const {btnStyle} = theme;
-
-  return (
-    <Container
-      maxW={'sm'}
-      shadow={'md'}
-      borderRadius={'lg'}
-      border={
-        colorMode === 'light' ? 'solid 1px #f4f6f8' : 'solid 1px #373737'
-      }>
-      <Box w={'100%'} p={0} textAlign={'center'} py={10} px={5}>
-        <Heading color={'blue.300'}>{userTonBalance} TON</Heading>
-        <Box py={5}>
-          <Text fontSize={'15px'} color={'gray.400'}>
-            Available in wallet
-          </Text>
-        </Box>
-        <Grid pos="relative" templateColumns={'repeat(2, 1fr)'} gap={6}>
-          <Button
-            {...(stakeDisabled === true
-              ? {...btnStyle.btnDisable({colorMode})}
-              : {...btnStyle.btnAble()})}
-            isDisabled={stakeDisabled}
-            fontSize={'14px'}
-            opacity={loading === true ? 0.5 : 1}
-            onClick={() => modalData('stake')}>
-            Stake
-          </Button>
-          <Button
-            {...(unstakeDisabled === true
-              ? {...btnStyle.btnDisable({colorMode})}
-              : {...btnStyle.btnAble()})}
-            isDisabled={unstakeDisabled}
-            fontSize={'14px'}
-            opacity={loading === true ? 0.5 : 1}
-            onClick={() => modalData('unstake')}>
-            Unstake
-          </Button>
-          <Button
-            {...(claimDisabled === true
-              ? {...btnStyle.btnDisable({colorMode})}
-              : {...btnStyle.btnAble()})}
-            isDisabled={claimDisabled}
-            fontSize={'14px'}
-            opacity={loading === true ? 0.5 : 1}
-            onClick={() => modalData('claim')}>
-            Claim
-          </Button>
-          {manageDisabled === true ? (
-            <Button
-              {...(endSaleBtnDisabled === true
-                ? {...btnStyle.btnDisable({colorMode})}
-                : {...btnStyle.btnAble()})}
-              isDisabled={endSaleBtnDisabled}
-              fontSize={'14px'}
-              opacity={loading === true ? 0.5 : 1}
-              onClick={() =>
-                data.miningEndTime !== undefined
-                  ? closeSale({
-                      userAddress: account,
-                      vaultContractAddress: data.vault,
-                      miningEndTime: data.miningEndTime,
-                      library: user.library,
-                      handleCloseModal: dispatch(closeModal()),
-                    })
-                  : null
-              }>
-              End Sale
-            </Button>
-          ) : (
-            <Button
-              {...(manageBtnDisabled === true
-                ? {...btnStyle.btnDisable({colorMode})}
-                : {...btnStyle.btnAble()})}
-              isDisabled={manageBtnDisabled}
-              fontSize={'14px'}
-              opacity={loading === true ? 0.5 : 1}
-              onClick={() => modalData('manage')}>
-              Manage
-            </Button>
-          )}
-
-          {loading === true ? (
-            <Flex
-              pos="absolute"
-              zIndex={100}
-              w="100%"
-              h="100%"
-              alignItems="cneter"
-              justifyContent="center">
-              <Center>
-                <LoadingComponent></LoadingComponent>
-              </Center>
-            </Flex>
-          ) : null}
-        </Grid>
-      </Box>
-    </Container>
   );
 };
 
@@ -394,7 +135,7 @@ export const Staking = () => {
     [],
   );
 
-  const GetTotalStaker = ({contractAddress, library}: any) => {
+  const GetTotalStaker = ({contractAddress, library, data}: any) => {
     const {colorMode} = useColorMode();
     const [totalStaker, setTotalStaker] = useState('-');
     const getlInfo = async () => {
@@ -403,7 +144,11 @@ export const Staking = () => {
     };
 
     useEffect(() => {
-      getlInfo();
+      if (user.address !== undefined) {
+        getlInfo();
+      } else {
+        setTotalStaker(data);
+      }
     }, []);
 
     return (
@@ -435,7 +180,9 @@ export const Staking = () => {
     };
 
     useEffect(() => {
-      getBalance();
+      if (user.address !== undefined) {
+        getBalance();
+      }
     }, [user]);
 
     if (user.address === undefined) {
@@ -520,7 +267,8 @@ export const Staking = () => {
             </Flex>
             <GetTotalStaker
               contractAddress={contractAddress}
-              library={library}></GetTotalStaker>
+              library={library}
+              data={data[row.id]?.totalStakers}></GetTotalStaker>
             <GetBalance
               title={'My staked'}
               contractAddress={contractAddress}
